assistant_git/requirements.txt.txt
pyttsx3
SpeechRecognition
keyboard
opencv-python
imdbpy
wolframalpha
pyautogui
python-decouple
python-dotenv
openai
requests
wikipedia
pywhatkit
Pillow
pygame
